feat(append-version): ZIP生成をやめて、バージョン追記スクリプトのみにする
